refactor: remove duplicated file; improve code style
